Modify the graph to use a custom chart module which displays dates
As per other modifications to the web interface I've had to copy a lot
of code from the original Mesa ChartModule but I can't see a way round
that.

In order to use Moment.js with Chart.js to enable a time series,
Moment.js needs to be loaded first; the simplest way to do this within
Mesa was to create a new minified JS file containing Moment (with
Moment Timezone) and Chart.js.